globals and parameters modules created. code structure changed
