+++ firmware/python/globals.py
@@ -1,0 +1,13 @@
+
+# Project:   berry_case
+# File:      globals.py
+# Autor:     Enrico Cirignaco
+# Created:   29.11.2020
+
+# Import modules
+#############################################################################
+
+menu_depth = 0
+main_menu_entry = 0
+network_menu_entry = 0
+system_info_menu_entry = 0
